+++ Test/ExternalAdviceTest/EmptyExternalAdvisedClass.cs
@@ -4,11 +4,13 @@
 // http://mradvice.arxone.com/
 // Released under MIT license http://opensource.org/licenses/mit-license.php
 #endregion
+
+using NUnit.Framework;
+
 namespace ExternalAdviceTest
 {
     using System.Reflection;
     using ExternalAdvices;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
 
     public class EmptyExternalAdvisedClass
     {
@@ -17,7 +19,7 @@ namespace ExternalAdviceTest
         {
             var thisMethod = MethodBase.GetCurrentMethod();
             F(thisMethod);
-            Assert.AreNotEqual("MethodTest", thisMethod.Name);
+            Assert.That(thisMethod.Name, Is.Not.EqualTo("MethodTest"));
         }
 
         [AddProperty]
@@ -25,7 +27,7 @@ namespace ExternalAdviceTest
         {
             var property = GetType().GetProperty("WeavingAdvisedMethodTest_Property");
             F(property);
-            Assert.IsNotNull(property);
+            Assert.That(property, Is.Not.Null);
         }
 
         void F(object o)