+++ Test/IntegrityTest/WildcardMatchTest.cs
@@ -1,86 +1,22 @@
-﻿#region Mr. Advice
-// Mr. Advice
-// A simple post build weaving package
-// http://mradvice.arxone.com/
-// Released under MIT license http://opensource.org/licenses/mit-license.php
-#endregion
+﻿using ArxOne.MrAdvice.Pointcut;
+using NUnit.Framework;
 
-namespace IntegrityTest
+[TestFixture]
+[Category("WildcardMatch")]
+public class WildcardMatchTest
 {
-    using ArxOne.MrAdvice.Pointcut;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-    [TestClass]
-    public class WildcardMatchTest
+    [TestCase("abc", "abc", ExpectedResult = true)]
+    [TestCase("def", "ghi", ExpectedResult = false)]
+    [TestCase("jk?", "jkl", ExpectedResult = true)]
+    [TestCase("m*", "mno", ExpectedResult = true)]
+    [TestCase("p@r", "pqr", ExpectedResult = true)]
+    [TestCase("p@", "p.r", ExpectedResult = false)]
+    [TestCase("p@", "pq.", ExpectedResult = false)]
+    [TestCase("st!", "stu", ExpectedResult = true)]
+    [TestCase("st!", "st.", ExpectedResult = false)]
+    [TestCase("*.<@>d@.@", "TestAsync.Program.Test.<<StartTask>b__0_0>d.MoveNext″", ExpectedResult = true)]
+    public bool Tests(string pattern, string input)
     {
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void SimpleEquals()
-        {
-            Assert.IsTrue(PointcutSelectorRule.WildcardMatch("abc", "abc"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void SimpleNotEquals()
-        {
-            Assert.IsFalse(PointcutSelectorRule.WildcardMatch("def", "ghi"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void QuestionMark()
-        {
-            Assert.IsTrue(PointcutSelectorRule.WildcardMatch("jk?", "jkl"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void Asterisk()
-        {
-            Assert.IsTrue(PointcutSelectorRule.WildcardMatch("m*", "mno"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void Arobase()
-        {
-            Assert.IsTrue(PointcutSelectorRule.WildcardMatch("p@r", "pqr"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void ArobaseAndDot()
-        {
-            Assert.IsFalse(PointcutSelectorRule.WildcardMatch("p@", "p.r"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void ArobaseAndFinalDot()
-        {
-            Assert.IsFalse(PointcutSelectorRule.WildcardMatch("p@", "pq."));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void ExclamationMark()
-        {
-            Assert.IsTrue(PointcutSelectorRule.WildcardMatch("st!", "stu"));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void ExclamationMarkAndDot()
-        {
-            Assert.IsFalse(PointcutSelectorRule.WildcardMatch("st!", "st."));
-        }
-
-        [TestMethod]
-        [TestCategory("WildcardMatch")]
-        public void MatchCase1()
-        {
-            Assert.IsTrue(PointcutSelectorRule.WildcardMatch("*.<@>d@.@", "TestAsync.Program.Test.<<StartTask>b__0_0>d.MoveNext″"));
-        }
+        return PointcutSelectorRule.WildcardMatch(pattern, input);
     }
 }