+++ Test/VBMethodLevelTest/EmptyAdvisedClass.vb
@@ -1,10 +1,12 @@
 ﻿Imports System.Reflection
 Imports VBMethodLevelTest.Advices
+Imports NUnit.Framework
 
 Public Class EmptyAdvisedClass
     <EmptyMethodAdvice()>
     Public Sub MethodTest()
         Dim thisMethod = MethodBase.GetCurrentMethod()
-        Assert.AreNotEqual("MethodTest", thisMethod.Name)
-        End Sub
+
+        Assert.That(thisMethod.Name <> "MethodTest")
+    End Sub
 End Class